% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencodersKeras.R
\name{autoencoder_keras}
\alias{autoencoder_keras}
\title{autoencoder_keras}
\usage{
autoencoder_keras(
  X,
  method = AUTOENCODER_METHODS,
  latent_dim = 3,
  time_window = 5,
  epochs = 200,
  batch_size = 32,
  optimizer_params = list(loss = "mean_squared_error", optimizer = "adam"),
  embedding_params = list(padNA = F, shift = time_window)
)
}
\arguments{
\item{X}{\itemize{
\item nxN matrix containing the N time series as columns, each one of length n time steps - Numeric
}}

\item{method}{\itemize{
\item Autoencoder type (from AUTOENCODER_METHODS list) - String
}}

\item{latent_dim}{\itemize{
\item Number of latent dimensions - Numeric
}}

\item{time_window}{\itemize{
\item Size of time window (in time steps) - Numeric
}}

\item{epochs}{\itemize{
\item Number of epochs required for training - Numeric
}}

\item{batch_size}{\itemize{
\item Batch size required for training - Numeric
}}

\item{optimizer_params}{\itemize{
\item Optimizer parameters for keras fit function
\itemize{
\item{\code{loss}: }{Loss function used for optimization (among those defined by Keras) - String}
\item{\code{optimizer}: }{Optimizer function used for optimization (among those defined by Keras) - String}
}
}}

\item{embedding_params}{\itemize{
\item Paramaters to control the embedding process
\itemize{
\item{\code{padNA}: }{Value used to indicate whether incomplete tensors should be dropped or filled with NAs - Boolean}
\item{\code{shift}: }{Number of time step that be skipped in order to determine the start of the following tensor - Numeric}
}
}}
}
\value{
List containing:
\itemize{
\item{\code{autoencoder}: }{Fitted autoencoder model (full model) - Keras object}
\item{\code{encoder}: }{Encoder part of the autoencoder model (Original space -> Latent Dimensions) - Keras object}
\item{\code{decoder}: }{Decoder part of the autoencoder model (Latent Dimensions -> Original space) - Keras object}
\item{\code{time_window}: }{Time window used for the embedding (to fit the autoencoder) - Numeric scalar}
\item{\code{train_history}: }{Object containing statistics on the training history - Keras object}
}
}
\description{
Wrapper function for all the different keras autoencoder implementations
}
