% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencodersKeras.R
\name{BaseRegularizedAutoencoder}
\alias{BaseRegularizedAutoencoder}
\title{BaseRegularizedAutoencoder}
\usage{
BaseRegularizedAutoencoder(X, latent_dim)
}
\arguments{
\item{X}{\itemize{
\item nxN matrix containing the N time series as columns, each one of length n time steps
}}

\item{latent_dim}{\itemize{
\item Number of latent dimensions (equivalent to the size of hidden layer) - numeric scalar
}}
}
\value{
List containing:
\itemize{
\item{\code{autoencoder}: }{Fitted autoencoder model (full model) - Keras object}
\item{\code{encoder}: }{Encoder part of the autoencoder model (Original space -> Latent Dimensions) - Keras object}
\item{\code{decoder}: }{Decoder part of the autoencoder model (Latent Dimensions -> Original space) - Keras object}
}
}
\description{
Auxiliary function defining a basic autoencoder model (1 dense hidden layer) with regularization
}
