% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencodersKeras.R
\name{Convolutional1DAutoencoder}
\alias{Convolutional1DAutoencoder}
\title{ConvolutionalAutoencoder}
\usage{
Convolutional1DAutoencoder(X, filters = c(32, 32, 16))
}
\arguments{
\item{X}{\itemize{
\item nxN matrix containing the N time series as columns, each one of length n time steps
}}

\item{filters}{\itemize{
\item Size of the the filters in the different layers - numeric vector
}}
}
\value{
List containing:
\itemize{
\item{\code{autoencoder}: }{Fitted autoencoder model (full model) - Keras object}
\item{\code{encoder}: }{Encoder part of the autoencoder model (Original space -> Latent Dimensions) - Keras object}
\item{\code{decoder}: }{Decoder part of the autoencoder model (Latent Dimensions -> Original space) - Keras object}
}
}
\description{
Auxiliary function defining a 1D convolutional autoencoder model (1 1DConv + 2x 1DConv+MaxPool)
}
