% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativePCA.R
\name{cor.prob}
\alias{cor.prob}
\title{cor.prob}
\usage{
cor.prob(X, dfr = nrow(X) - 2)
}
\arguments{
\item{X}{\itemize{
\item n x n correlation matrix - numeric
}}

\item{dfr}{\itemize{
\item Degrees of freedom - numeric
}}
}
\value{
\itemize{
\item \code{R} - Matrix containing the probabilities of statistical significance of the different correlations
}
}
\description{
Auxiliary function to perform statistical testing of the correlation levels in a correlation matrix
}
