% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionalityReduction.R
\name{decode}
\alias{decode}
\title{decode}
\usage{
decode(model, Z_hat)
}
\arguments{
\item{model}{\itemize{
\item Autoencoder model to employ to perfom the decoding
}}

\item{Z_hat}{\itemize{
\item h x k matrix containing the h-step ahead forecast for the k input series - Numeric
}}
}
\value{
\code{X_hat} - h x N matrix containing the h-step ahead forecast for the N original series
}
\description{
Auxilary function to perform dimensionality increase with autoencoder techniques (only for those non based on Keras).
}
