% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencodersKeras.R
\name{DeepAutoencoder}
\alias{DeepAutoencoder}
\title{DeepAutoencoder}
\usage{
DeepAutoencoder(X, latent_dim = c(10, 5, 2))
}
\arguments{
\item{X}{\itemize{
\item nxN matrix containing the N time series as columns, each one of length n time steps
}}

\item{latent_dim}{\itemize{
\item Number of latent dimensions (equivalent to the sizes of hidden layers) - numeric vector
}}
}
\value{
List containing:
\itemize{
\item{\code{autoencoder}: }{Fitted autoencoder model (full model) - Keras object}
\item{\code{encoder}: }{Encoder part of the autoencoder model (Original space -> Latent Dimensions) - Keras object}
\item{\code{decoder}: }{Decoder part of the autoencoder model (Latent Dimensions -> Original space) - Keras object}
}
}
\description{
Auxiliary function defining a deep autoencoder model (3 dense hidden layer)
}
