% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencodersKeras.R
\name{DeepGRUAutoencoder}
\alias{DeepGRUAutoencoder}
\title{DeepGRUAutoencoder}
\usage{
DeepGRUAutoencoder(X, latent_dim = c(10, 5), time_window = 5)
}
\arguments{
\item{X}{\itemize{
\item nxN matrix containing the N time series as columns, each one of length n time steps
}}

\item{latent_dim}{\itemize{
\item Number of GRU cells in the hidden layers - numeric vector
}}

\item{time_window}{\itemize{
\item Size of time window (in time steps) for embedding - numeric scalar
}}
}
\value{
List containing:
\itemize{
\item{\code{autoencoder}: }{Fitted autoencoder model (full model) - Keras object}
\item{\code{encoder}: }{Encoder part of the autoencoder model (Original space -> Latent Dimensions) - Keras object}
\item{\code{decoder}: }{Decoder part of the autoencoder model (Latent Dimensions -> Original space) - Keras object}
}
}
\description{
Auxiliary function defining a GRU autoencoder model with two hidden layers
}
