% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFML.R
\name{DFML}
\alias{DFML}
\title{DFML}
\usage{
DFML(
  X,
  dimensionality_method = DIMENSIONALITY_METHODS,
  forecast_family = FORECAST_FAMILY,
  dimensionality_parameters,
  forecast_parameters,
  components,
  h
)
}
\arguments{
\item{X}{\itemize{
\item nxN matrix containing the N time series as columns, each one of length n time steps
}}

\item{dimensionality_method}{\itemize{
\item Dimensionality reduction method to employ - String among those defined in DIMENSIONALITY_METHOD
}}

\item{forecast_family}{\itemize{
\item Forecasting family of method to employ - String among those defined in FORECAST_FAMILY
}}

\item{dimensionality_parameters}{\itemize{
\item Parameters to be passed to the dimensionalityReduction function - List
}}

\item{forecast_parameters}{\itemize{
\item Parameters to be passed to the forecaster function - List
}}

\item{components}{\itemize{
\item Number of desired factors - numeric scalar
}}

\item{h}{\itemize{
\item Forecasting horizon - numeric scalar
}}
}
\value{
List containing:
\itemize{
\item{\code{X_hat}: }{hxN matrix containing the forecasts of N time series as columns, each one of length h time steps}
\item{\code{Model_dim}: }{Dimensionality reduction model as produced by dimensionality_reduction function}
\item{\code{Time_dim}: }{Computational time required to run the dimensionality reduction model - numeric scalar}
\item{\code{Time_forecast}: }{Computational time required to run the forecasting model - numeric scalar}
}
}
\description{
Core function implementing the EDFML technique.
Takes the input data X, and passes it through the dimensionality reduction step (via the dimensionalityReduction fucntion) to obtaing the dynamic factors Z.
Factors are then forecast (via the forecaster function), in order to obtain the factor forecasts (Z_hat).
Finally, factor forecast are transformed to forecast in the original space via an inverse dimensionality reduction (dimensionality increase).
}
\examples{
#See tests/testthat directory on https://github.com/jdestefani/ExtendedDFML
}
