% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionalityReduction.R
\name{dimensionalityIncrease}
\alias{dimensionalityIncrease}
\title{dimensionalityIncrease}
\usage{
dimensionalityIncrease(
  Z_hat,
  family = DIMENSIONALITY_METHODS,
  model = NULL,
  Z = NULL,
  params = NULL
)
}
\arguments{
\item{Z_hat}{\itemize{
\item h x k matrix containing the h-step ahead forecast for the k input series - Numeric
}}

\item{family}{\itemize{
\item Family of methods to use - String among those defined in DIMENSIONALITY_METHODS
}}

\item{model}{\itemize{
\item An (optional) pre-existing model. If NULL it will be computed by the function.
}}

\item{Z}{\itemize{
\item n x k matrix contaning the k factor generated by the dimensionality reduction module - Numeric
}}

\item{params}{\itemize{
\item Additional parameters to be passed to the forecaster function - List
}}
}
\value{
List containing:
\itemize{
\item{\code{X_hat}: }{h x N matrix containing the h-step ahead forecast for the N original series}
\item{\code{time_dim}: }{Computational time (in s) required to run the dimensionality increase - numeric scalar}
}
}
\description{
Function implementing the dimensionality increase module in the DFML architecture.
The function implements a batch dimensionality increase.
}
\examples{
#See tests/testthat directory on https://github.com/jdestefani/ExtendedDFML
}
