% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecaster.R
\name{forecaster}
\alias{forecaster}
\title{forecaster}
\usage{
forecaster(Z, family = FORECAST_FAMILY, parameters = NULL, h)
}
\arguments{
\item{Z}{\itemize{
\item nxk matrix containing the k time series as columns, each one of length n time steps
}}

\item{family}{\itemize{
\item Forecasting family of method to employ - String among those defined in FORECAST_FAMILY
}}

\item{parameters}{\itemize{
\item Parameters to be passed to the forecaster function - List\preformatted{              The method name is passed through the \code{parameters$method} and should be one among those defined in \code{MULTISTEPAHEAD_METHODS}, \code{M4_METHODS}, \code{GRADIENT_BOOSTING_METHODS} 
              
              For the different methods family, at least the embedding order/model order \code{m} is required.
              
              For additional parameters:
              \\itemize\{
                \\item\{\code{MULTISTEPAHEAD_METHODS}: \}\{See \pkg{gbcode::multipleStepAhead} documentation for the role of the different parameters\}
                \\item\{\code{M4_METHODS}: \}\{No specific parameters required\}
                \\item\{\code{GRADIENT_BOOSTING_METHODS}: \}\{ 
                       \\itemize\{
                          \\item\{\code{multistep_method:}{Multistep method to be chosen between \code{direct} and \code{recursive}}\}
                          \\item\{\code{forecasting_method:}{Forecasting method to be chosen between \code{lightgbm} and \code{xgboost}}\}
                        \}
                       See \pkg{lightgbm} and \pkg{xgboost} documentation for the role of the different parameters\}
              \}
}
}}

\item{h}{\itemize{
\item Forecasting horizon - numeric scalar
}}
}
\value{
List containing:
\itemize{
\item{\code{Z_hat}: }{h x k matrix containing the h-step ahead forecast for the k input series}
\item{\code{time_forecast}: }{Computational time required to run the forecasting model - numeric scalar}
}
}
\description{
Function implementing the forecasting module in the DFML architecture
}
\examples{
#See tests/testthat directory on https://github.com/jdestefani/ExtendedDFML
}
