% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientBoosting.R
\name{gradientBoostingForecaster}
\alias{gradientBoostingForecaster}
\title{gradientBoostingForecaster}
\usage{
gradientBoostingForecaster(
  ts,
  embedding,
  horizon,
  delay = 0,
  multistep_method = c("recursive", "direct"),
  forecasting_method = c("lightgbm", "xgboost"),
  forecasting_params = NULL
)
}
\arguments{
\item{ts}{\itemize{
\item Input time series as a vector
}}

\item{embedding}{\itemize{
\item Embedding order
}}

\item{horizon}{\itemize{
\item Forecasting horizon
}}

\item{delay}{\itemize{
\item Delay for forecasting
}}

\item{multistep_method}{\itemize{
\item Multistep method (to be chosen among "recursive" and "direct")
}}

\item{forecasting_method}{\itemize{
\item Gradient boosting forecasting method (to be chosen among "lightgbm" and "xgboost")
}}

\item{forecasting_params}{\itemize{
\item Parameters to be passed to the forecaster function - List
\itemize{
\item{\code{multistep_method:}{Multistep method to be chosen between \code{direct} and \code{recursive}}}
\item{\code{forecasting_method:}{Forecasting method to be chosen between \code{lightgbm} and \code{xgboost}}}
}\preformatted{                       See \pkg{lightgbm} and \pkg{xgboost} documentation for the role of the different parameters
}
}}
}
\value{
h-step forecast of ts using the chosen multistep_method and forecasting method, with the given embedding order
}
\description{
gradientBoostingForecaster
}
