% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionalityReduction.R
\name{GramSchmidtOrthogonalization}
\alias{GramSchmidtOrthogonalization}
\title{GramSchmidtOrtogonalization}
\usage{
GramSchmidtOrthogonalization(Z, Z_hat, k)
}
\arguments{
\item{Z}{\itemize{
\item n x k matrix contaning the k factor generated by the dimensionality reduction module - Numeric
}}

\item{Z_hat}{\itemize{
\item h x k matrix containing the h-step ahead forecast for the k input series - Numeric
}}

\item{k}{\itemize{
\item Number of factors for the orthogonalization - Numeric
}}
}
\value{
\code{U} - (n+h) x k matrix containing the orthogonalized factors
}
\description{
Helper function to implement Gram-Schmidt Ortogonalization (applied to the factor forecast)
}
