% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencodersKeras.R
\name{incremental_autoencoder_keras_update}
\alias{incremental_autoencoder_keras_update}
\title{incremental_autoencoder_keras}
\usage{
incremental_autoencoder_keras_update(
  X_update,
  method,
  model,
  time_window = 5,
  epochs = 200,
  batch_size = 32
)
}
\arguments{
\item{X_update}{\itemize{
\item nxN matrix containing the N time series as columns, each one of length n time steps used to perform the model update
}}

\item{method}{\itemize{
\item Autoencoder type (from AUTOENCODER_METHODS list)
}}

\item{model}{\itemize{
\item Pre-trained autoencoder model
}}

\item{time_window}{\itemize{
\item Size of time window (in time steps)
}}

\item{epochs}{\itemize{
\item Number of epochs required for training
}}

\item{batch_size}{\itemize{
\item Batch size required for training
}}
}
\value{
List containing:
\itemize{
\item{\code{autoencoder}: }{Fitted autoencoder model (full model) - Keras object}
\item{\code{encoder}: }{Encoder part of the autoencoder model (Original space -> Latent Dimensions) - Keras object}
\item{\code{decoder}: }{Decoder part of the autoencoder model (Latent Dimensions -> Original space) - Keras object}
\item{\code{time_window}: }{Time window used for the embedding (to fit the autoencoder) - Numeric scalar}
\item{\code{train_history}: }{Object containing statistics on the training history - Keras object}
}
}
\description{
Wrapper function for an incremental fitting of an autoencoder model
}
