% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionalityReduction.R
\name{incrementalDimensionalityReduction}
\alias{incrementalDimensionalityReduction}
\title{incrementalDimensionalityReduction}
\usage{
incrementalDimensionalityReduction(
  X,
  init_sample_percentage,
  k,
  family = DIMENSIONALITY_METHODS,
  params = list(model = NULL)
)
}
\arguments{
\item{X}{\itemize{
\item n x N matrix containing the N time series as columns, each one of length n time steps
}}

\item{init_sample_percentage}{\itemize{
\item Numeric value indicating the percentage of \code{X} to be used for model fit - Numeric
}}

\item{k}{\itemize{
\item Number of desired factors as output - Numeric
}}

\item{family}{\itemize{
\item Family of methods to use - String among those defined in DIMENSIONALITY_METHODS
}}

\item{params}{\itemize{
\item Parameters to be passed to the function - List\preformatted{          An (optional) pre-existing model is passed through the \code{parameters$model}.
          If NULL it will be computed by the function.
}
}}
}
\value{
List containing:
\itemize{
\item{\code{Z}: }{n x k matrix contaning the k factor generated by the dimensionality reduction module}
\item{\code{time_dim}: }{Computational time (in s) required to fit the dimensionality model - numeric scalar}
\item{\code{model}: }{Fitted model constructed via the function}
}
}
\description{
Function implementing the dimensionality reduction module in the DFML architecture.
The function implements an incremental dimensionality reduction model with a rolling approach.
A percentage of \code{X}, defined by \code{init_sample_percentage} is used to initialize the dimensionality reduction technique.
The remaining part of \code{X} is used to incrementally fit the dimensionality reduction model
}
\examples{
#See tests/testthat directory on https://github.com/jdestefani/ExtendedDFML
}
