% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativePCA.R
\name{incrementalPCA}
\alias{incrementalPCA}
\title{incrementalPCA}
\usage{
incrementalPCA(
  X,
  xbar = NULL,
  base_pca_values,
  base_pca_vectors,
  pca_type = INCREMENTAL_PCA_METHODS,
  forgetting_factor = NA,
  components = 2,
  start_value,
  end_value
)
}
\arguments{
\item{X}{\itemize{
\item n x N matrix containing the N time series as columns, each one of length n time steps
}}

\item{xbar}{\itemize{
\item 1 x N matrix containing the means of the N input time series
}}

\item{base_pca_values}{\itemize{
\item Matrix containing the PCA eigenvalues to use for the initialization
}}

\item{base_pca_vectors}{\itemize{
\item Matrix containing the PCA eigenvectors to use for the initialization
}}

\item{pca_type}{\itemize{
\item Incremental PCA method to employ - String among those defined in \code{INCREMENTAL_PCA_METHODS}
}}

\item{forgetting_factor}{\itemize{
\item Forgetting factor to employ in the incremental PCA - Numeric
}}

\item{components}{\itemize{
\item Desired number of PCA components - Numeric
}}

\item{start_value}{\itemize{
\item Index of the first value in the dataset to be used for incremental PCA - Numeric
}}

\item{end_value}{\itemize{
\item Index of the last value in the dataset to be used for incremental PCA - Numeric
}}
}
\value{
List containing:
\itemize{
\code{pca} - List containing:
\itemize{
\item{\code{values}: }{Matrix containing the eigenvalues associated to PCA}
\item{\code{vectors}: }{Matrix containing the eigenvectors associated to PCA as columns}
}
\item{\code{x_bar}: }{1 x N matrix containing the means of the N input time series}
}
}
\description{
Wrapper function for an iterative PCA method (onlinePCA::incRpca)
}
