% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativePCA.R
\name{incrementalpca_wrapper}
\alias{incrementalpca_wrapper}
\title{incrementalpca_wrapper}
\usage{
incrementalpca_wrapper(
  X,
  components = 10,
  init_sample_percentage = 0.5,
  stop_iterative = FALSE,
  centered = TRUE
)
}
\arguments{
\item{X}{\itemize{
\item Input dataset - Numeric matrix
}}

\item{components}{\itemize{
\item Desired number of PCA components - Numeric
}}

\item{init_sample_percentage}{\itemize{
\item Percentage of the input dataset to be used as initialisation for iterative PCA - numeric
}}

\item{stop_iterative}{\itemize{
\item Boolean flag to select whether the iterative part would be performed or not - boolean
}}

\item{centered}{\itemize{
\item Boolean flag to selected whether a centering of the dataset would be performed or not - boolean
}}
}
\value{
\code{pca} - List containing:
\itemize{
\item{\code{values}: }{Matrix containing the eigenvalues associated to PCA}
\item{\code{vectors}: }{Matrix containing the eigenvectors associated to PCA as columns}
}
}
\description{
Wrapper function for the iterative PCA method (onlinePCA::incRpca).
}
