% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativePCA.R
\name{iterativePCASingleStep}
\alias{iterativePCASingleStep}
\title{iterativePCASingleStep}
\usage{
iterativePCASingleStep(
  X_i,
  i,
  pca_type,
  xbar,
  pca,
  forgetting_factor = NA,
  components
)
}
\arguments{
\item{X_i}{\itemize{
\item 1 x N matrix containing the sample of the original dataset to use for the update - Numeric
}}

\item{i}{\itemize{
\item Index of the sample - numeric
}}

\item{pca_type}{\itemize{
\item Incremental PCA method to employ - String among those defined in \code{INCREMENTAL_PCA_METHODS}
}}

\item{xbar}{\itemize{
\item 1 x N matrix containing the means of the N input time series
}}

\item{pca}{\itemize{
\item List containing:
\itemize{
\item{\code{values}: }{Matrix containing the eigenvalues associated to PCA}
\item{\code{vectors}: }{Matrix containing the eigenvectors associated to PCA as columns}
}
}}

\item{forgetting_factor}{\itemize{
\item Forgetting factor to employ in the incremental PCA - Numeric
}}

\item{components}{\itemize{
\item Desired number of PCA components - Numeric
}}
}
\value{
List containing:
\itemize{
\code{pca} - List containing:
\itemize{
\item{\code{values}: }{Matrix containing the eigenvalues associated to PCA}
\item{\code{vectors}: }{Matrix containing the eigenvectors associated to PCA as columns}
}
\item{\code{x_bar}: }{1 x N matrix containing the means of the N input time series}
}
}
\description{
Wrapper function implementing a single update in the iterative PCA procedure
}
