% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencodersKeras.R
\name{learningRateDecay}
\alias{learningRateDecay}
\title{learningRateDecay}
\usage{
learningRateDecay(epoch)
}
\arguments{
\item{epoch}{\itemize{
\item Epoch parameter
}}
}
\description{
Auxiliary function to be passed to Keras callbacks in order to perform learning rate decay
}
