% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencodersKeras.R
\name{LSTMAutoencoder}
\alias{LSTMAutoencoder}
\title{LSTMAutoencoder}
\usage{
LSTMAutoencoder(X, latent_dim = 3, time_window = 5)
}
\arguments{
\item{X}{\itemize{
\item nxN matrix containing the N time series as columns, each one of length n time steps
}}

\item{latent_dim}{\itemize{
\item Number of LSTM cells in the hidden layer - numeric scalar
}}

\item{time_window}{\itemize{
\item Size of time window (in time steps) for embedding - numeric scalar
}}
}
\value{
List containing:
\itemize{
\item{\code{autoencoder}: }{Fitted autoencoder model (full model) - Keras object}
\item{\code{encoder}: }{Encoder part of the autoencoder model (Original space -> Latent Dimensions) - Keras object}
\item{\code{decoder}: }{Decoder part of the autoencoder model (Latent Dimensions -> Original space) - Keras object}
}
}
\description{
Auxiliary function defining a LSTM autoencoder model with one hidden layer
}
