% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencodersKeras.R
\name{LSTMConvolutional1DAutoencoder}
\alias{LSTMConvolutional1DAutoencoder}
\title{LSTMConvolutional1DAutoencoder}
\usage{
LSTMConvolutional1DAutoencoder(X, latent_dim = c(32, 32, 16))
}
\arguments{
\item{X}{\itemize{
\item nxN matrix containing the N time series as columns, each one of length n time steps
}}

\item{latent_dim}{\itemize{
\item Number of filters (1st value) / LSTM cells (2nd/3rd values) in the hidden layers - numeric vector
}}
}
\value{
List containing:
\itemize{
\item{\code{autoencoder}: }{Fitted autoencoder model (full model) - Keras object}
\item{\code{encoder}: }{Encoder part of the autoencoder model (Original space -> Latent Dimensions) - Keras object}
\item{\code{decoder}: }{Decoder part of the autoencoder model (Latent Dimensions -> Original space) - Keras object}
}
}
\description{
Auxiliary function defining a convolutional + LSTM autoencoder model with one convolutional and two hidden layers
}
