% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingFunctions.R
\name{matrix2Tensor3D}
\alias{matrix2Tensor3D}
\title{matrix2Tensor3D}
\usage{
matrix2Tensor3D(X, time_window, shift = 1, padNA = FALSE)
}
\arguments{
\item{X}{\itemize{
\item Input matrix
}}

\item{time_window}{\itemize{
\item Integer specifying the size of the time window (aka model order, 2nd dimension in tensor).
}}

\item{shift}{\itemize{
\item Integer specifying how far the sliding window should be shifted in order to create the tensor (N.B. shift < time_window implies overlapping data)
}}

\item{padNA}{\itemize{
\item Boolean specifying whether NAs should be used to pad tensors in case the input matrix size is not multiple of the time_window.
}}
}
\value{
3D Tensor (Samples,Time Window,Features) containing the embedded form of X (the second dimension is fixed to time_window)
}
\description{
Auxiliary function to embed a 2D matrix into a 3D tensor.
}
\examples{
X <- EuStockMarkets
time_window <- 5
X_windowed <- matrix2Tensor3D(X,time_window,time_window)
}
