% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencodersKeras.R
\name{plotLoss}
\alias{plotLoss}
\title{plotLoss}
\usage{
plotLoss(model)
}
\arguments{
\item{model}{\itemize{
\item Keras model containing the history of the loss function to plot
}}
}
\description{
Auxiliary function to plot the evolution of the loss function
}
