% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddingFunctions.R
\name{tensor3D2matrix}
\alias{tensor3D2matrix}
\title{tensor3D2matrix}
\usage{
tensor3D2matrix(X, shift = 1)
}
\arguments{
\item{X}{\itemize{
\item Input 3D Tensor (Samples,Time Window,Features) containing the embedded form of X
}}

\item{shift}{\itemize{
\item Integer specifying the shift value used to construct the tensor
}}
}
\value{
2D Matrix (Samples,Features) obtained by "unrolling" the tensor
}
\description{
Auxiliary function to flatten 3D tensor into a 2D matrix, discarding overlapping data.
}
\examples{
X <- EuStockMarkets
time_window <- 5
X_windowed <- matrix2Tensor3D(X,time_window,time_window)
X_reconstructed <- tensor3D2matrix(X_windowed,time_window)
}
