% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativePCA.R
\name{vectorNormalize}
\alias{vectorNormalize}
\title{vectorNormalize
Normalize the vector passed as parameter}
\usage{
vectorNormalize(u)
}
\arguments{
\item{u}{\itemize{
\item Vector to normalize
}}
}
\value{
Normalized vector
}
\description{
vectorNormalize
Normalize the vector passed as parameter
}
