% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionalityReduction.R, R/iterativePCA.R
\name{vectorProject}
\alias{vectorProject}
\title{vectorProject}
\usage{
vectorProject(u, v)

vectorProject(u, v)
}
\arguments{
\item{u}{\itemize{
\item Vector to project
}}

\item{v}{\itemize{
\item Vector to project onto
}}
}
\value{
\itemize{
\item Numeric vector representing the result of the projection
}

Projection of u onto v
}
\description{
Auxiliary function to compute vector projection of vector u onto vector v.
}
