% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientBoosting.R
\name{xgboostRecursive}
\alias{xgboostRecursive}
\title{xgboostRecursive}
\usage{
xgboostRecursive(ts_embed, horizon, params = NULL)
}
\arguments{
\item{ts_embed}{\itemize{
\item Embedded time series, in the inp,out format of gbcode::makeEmbedded
}}

\item{horizon}{\itemize{
\item Forecasting horizon
}}

\item{params}{\itemize{
\item Additional (optional) parameters for XGBoost
}}
}
\value{
h-step forecast of ts_embed using xgboost and recursive multistep-ahead forecasting technique
}
\description{
Wrapper function for XGboost forecasting technique with recursive multistep ahead technique
}
