% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativePCA.R
\name{vectorProject}
\alias{vectorProject}
\title{vectorProject
Compute the vector projection of the vector u passed as parameters onto the vector v}
\usage{
vectorProject(u, v)
}
\arguments{
\item{u}{\itemize{
\item Vector to project
}}

\item{v}{\itemize{
\item Vector to project onto
}}
}
\value{
Projection of u onto v
}
\description{
vectorProject
Compute the vector projection of the vector u passed as parameters onto the vector v
}
